+++ Steps/AboutUsSteps.cs
@@ -1,5 +1,5 @@
 using Microsoft.Playwright;
-using playwright_c_.pageObject;
+using playwright_c_.PageObject;
 using PlaywrightTests.Utils;
 using TechTalk.SpecFlow;
 using static Microsoft.Playwright.Assertions;
